requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-anthropic>=0.1.0
chromadb>=0.4.0
gitpython>=3.1.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
